+++ Guinealogia/Objects/Quiz Contract.swift
