+++ Guinealogia/Objects/QuizState.swift
