+++ Guinealogia/Views/CodigoQR.swift
