+++ Guinealogia/Views/Modo Libre/ MenuModoLibre.swift
@@ -119,6 +119,7 @@ struct MenuModoLibre: View {
                 Spacer()
             }
         }
+        .navigationBarBackButtonHidden(true)
         .onAppear {
             loadPlayerName()
             loadHighScore()
