+++ Guinealogia/Views/Modo Libre/ResultadoView.swift
