+++ Guinealogia/Views/ModoCompeticion/ResultadoCompeticion.swift
